refactor(instrumentation-amqplib): add code comments for vendoring
